@@ -149,15 +149,13 @@
     "@types/mocha": "^10.0.1",
     "@types/ncp": "^2.0.5",
     "@types/node-fetch": "^2.5.3",
-    "@types/rmfr": "^2.0.1",
     "@types/semver": "^7.1.0",
     "@types/serialport": "^8.0.2",
     "@types/split": "^1.0.0",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/unzipper": "^0.10.5",
     "@types/uuid": "^8.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
+    "@typescript-eslint/eslint-plugin": "^8.16.0",
     "chai": "^4.0.0",
     "chai-json-equal": "0.0.1",
     "chai-things": "^0.2.0",
@@ -169,11 +167,10 @@
     "lint-staged": "^10.0.3",
     "mocha": "^10.2.0",
     "prettier": "^2.7.1",
-    "rimraf": "^3.0.2",
-    "rmfr": "^2.0.0",
+    "rimraf": "^6.0.1",
     "ts-node": "^10.9.1",
-    "typedoc": "^0.23.23",
-    "typescript": "^4.8.3"
+    "typedoc": "^0.26.11",
+    "typescript": "5.6.x"
   },
   "funding": "https://opencollective.com/signalk"
 }
